+++ .github/workflows/wheels.yml
@@ -3,6 +3,7 @@ on:
   push:
     branches:
       - main
+      - develop
   release:
     types:
       - created
@@ -28,7 +29,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-20.04, macOS-10.15]
+        os: [ubuntu-20.04, macOS-10.15, windows-2019]
 
     steps:
       - uses: actions/checkout@v2
@@ -43,7 +44,7 @@ jobs:
         run: python -m cibuildwheel --output-dir wheelhouse
         env:
           CIBW_BUILD: "cp36-* cp37-* cp38-* cp39-*"
-          CIBW_ENVIRONMENT: "CFLAGS='-march=core-avx-i' CXXFLAGS='-march=core-avx-i'"
+          CIBW_ENVIRONMENT: "CFLAGS='-march=core-avx-i' CXXFLAGS='-march=core-avx-i' CL='/arch:AVX'"
           CIBW_TEST_COMMAND: pytest {project}/tests
           CIBW_TEST_REQUIRES: pytest
